+++ src/components/dashboard/Developer/DeveloperDashboard.js
@@ -4,7 +4,7 @@ import NotificationCenter from "../NotificationCenter/NotificationContainer";
 import styled from "styled-components";
 import { connect } from "react-redux";
 import { Route, Redirect } from "react-router-dom";
-import ProjectIntroduction from "../Developer/DevProjectList/ProjectIntroduction"
+import HelloUser from "./Dashboard/Hello User/HelloUser"
 import ProjectsList from "./Dashboard/Projects Loader/ProjectsList";
 
 const DashboardContentContainer = styled.div`
@@ -55,7 +55,7 @@ class DeveloperDashboard extends Component {
           render={() => {
             return (
               <div className="developer-home">
-                <ProjectIntroduction />
+                <HelloUser />
                 <div className="project-board-and-notifications">
                   <ProjectBoard {...this.props} />
                   {this.renderNotificationAndStatement()}
